refactor: TitleBox 컴포넌트가 title, description prop을 받도록 수정
