+++ src/pages/juniorPromise/constants/constants.ts
@@ -2,12 +2,10 @@ export const TIME_SELECTION_BUTTON = ['첫 번째 일정 선택하기', '두 번
 
 export const SELECT_JUNIOR_TITLE = [
   {
-    id: 0,
     title: '선약 일정을 선택해주세요',
     description: '제안해 주신 3가지 일정 중 선배가 하나를 선택하면 확정돼요',
   },
   {
-    id: 1,
     title: '어떤 고민이 있나요?',
     description: '선배가 미리 확인하고 답변을 준비하도록 전달해요',
   },
